check if database exists before dropping it to improve performance when inserting the data the first time

+++ many_databases_test.js
@@ -29,7 +29,9 @@ function insert_and_form_operations (mongodb) {
 
         var db = mongodb.getSisterDB('boom-' + i);
         // drop every database 'boom-*'
-        db.dropDatabase();
+        if (db.getCollectionNames().length !== 0 ) {
+            db.dropDatabase();
+        }
 
         for (var y = 0; y < numCols; y++) {
             var bulkInsert = [];
